Move content of env_val.rs to convert.rs
